backend/copy-trading/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
redis==5.0.1
kafka-python==2.0.2
